lol been a while since i committed

+++ src/bots/timer/main.py
@@ -7,44 +7,47 @@ client = None
 
 class TimerClient(BotClient):
   def __init__(self):
-    BotClient.__init__(self, "timer")
+    BotClient.__init__(self, "")
     self.name = "timer"
 
 client = TimerClient()
 
-@client.command("Testing Commands", ["test"], "test", "Test the Timer bot")
-async def command_test(command, message):
-  await send(message, "Test success!", reaction = "check")
+timers = {}
 
-@client.command("Timer Commands", ["start"], "start", "Start a standard 5-minute BP timer")
+@client.command("Timer Commands", ["-start"], "-start", "Start a 5-minute BP timer; 30 seconds protected time, 15 seconds grace.")
 async def command_start(command, message):
+  cid = message.channel.id
+  mid = message.id
+  timers[cid] = timers.get(cid, []) + [mid]
   replies = []
   replies.append(await send(message, "5-minute timer started!", reaction = "check"))
   await asyncio.sleep(30)
-  replies.append(await send(message, "Protected time over!"))
+  if mid in timers[cid]: replies.append(await send(message, "Protected time over!"))
   await asyncio.sleep(30)
-  replies.append(await send(message, "4 minutes remaining!"))
+  if mid in timers[cid]: replies.append(await send(message, "4 minutes remaining!"))
   await asyncio.sleep(60)
-  replies.append(await send(message, "3 minutes remaining!"))
+  if mid in timers[cid]: replies.append(await send(message, "3 minutes remaining!"))
   await asyncio.sleep(60)
-  replies.append(await send(message, "2 minutes remaining!"))
+  if mid in timers[cid]: replies.append(await send(message, "2 minutes remaining!"))
   await asyncio.sleep(60)
-  replies.append(await send(message, "1 minute remaining!"))
+  if mid in timers[cid]: replies.append(await send(message, "1 minute remaining!"))
   await asyncio.sleep(30)
-  replies.append(await send(message, "Protected time!"))
+  if mid in timers[cid]: replies.append(await send(message, "Protected time!"))
   await asyncio.sleep(30)
-  replies.append(await send(message, "15-second grace period!"))
+  if mid in timers[cid]: replies.append(await send(message, "15-second grace period!"))
   await asyncio.sleep(5)
-  replies.append(await send(message, "10 seconds!"))
+  if mid in timers[cid]: replies.append(await send(message, "10 seconds!"))
   await asyncio.sleep(5)
-  replies.append(await send(message, "5 seconds!"))
+  if mid in timers[cid]: replies.append(await send(message, "5 seconds!"))
   await asyncio.sleep(5)
-  replies.append(await send(message, "Time's Up!"))
+  if mid in timers[cid]: replies.append(await send(message, "Time's Up!"))
   await asyncio.sleep(5)
-  await send(message, "[a 5-minute timer was run]")
-  await message.channel.delete_messages(replies)
+  if mid in timers[cid]: await send(message, "[a 5-minute timer was run]")
+  if mid in timers[cid]: await message.channel.delete_messages(replies)
+
+@client.command("Timer Commands", ["-stop"], "-stop", "Stops all timers in this channel.")
+async def command_stop(command, message):
+  timers[message.channel.id] = []
+  await send(message, "All timers in this channel stopped!", reaction = "check")
 
 set_client(client)
-
-def start():
-  client.run(config["discord-tokens"]["timer"])
